requirements.txt
django==1.10.2
Pillow
django-recaptcha
